Fix many things, pool use, parallelization, VAF

+++ services/variantannotation/vannot/cli/dockerfile/bin/dejavu_processing.py
@@ -7,16 +7,22 @@ import subprocess
 import commons
 import time
 from time import sleep
+from multiprocessing import Pool
 
 import howard_launcher
 import howard_processing
 
-
+def divide_memory(threads: str, memory: str) -> str:
+    suffix = memory[-1]
+    memory = int(memory[:-1])
+    memory = memory/int(threads)
+    return str(memory) + suffix
+    
 def convert_vcf_parquet(run_informations, args):
     threads = commons.get_threads("threads_dejavu")
     memory = commons.get_memory("memory_dejavu")
     run_dict = {}
-
+    
     log.info(
         f"Generating new dejavu database for {run_informations["run_platform_application"]}"
     )
@@ -34,11 +40,34 @@ def convert_vcf_parquet(run_informations, args):
         os.makedirs(run_informations["tmp_analysis_folder"], 0o775)
 
     if "run" in args:
-        vcf_files = glob.glob(osj(run_informations["archives_run_folder"], "*.vcf.gz"))
+        vcf_files_archives = glob.glob(osj(run_informations["archives_run_folder"], "*.vcf.gz"))
     elif "dejavu" in args:
-        vcf_files = glob.glob(
+        vcf_files_archives = glob.glob(
             osj(run_informations["archives_project_folder"], "VCF", "*", "*.vcf.gz")
         )
+
+    for vcf_file in vcf_files_archives:
+        output = subprocess.check_output(f'zgrep -v \"#\" {vcf_file} | wc -l', shell=True, text=True)
+        if output == 0:
+            vcf_files_archives.remove(vcf_file)
+
+    for vcf_file in vcf_files_archives:
+        vcf_file_rundir = vcf_file.split("/")[-2]
+        new_tmp = osj(run_informations["tmp_analysis_folder"], vcf_file_rundir)
+        if not os.path.isdir(new_tmp):
+            os.makedirs(new_tmp)
+        subprocess.call(
+            ["rsync", "-rvt", vcf_file, new_tmp],
+            universal_newlines=True,
+        )
+    vcf_files = glob.glob(
+        osj(run_informations["tmp_analysis_folder"], "*", "*.vcf.gz")
+    )
+
+    if "dejavu" in args:
+        for vcf_file in vcf_files:
+            howard_processing.unmerge_vcf(vcf_file, run_informations)
+        vcf_files = glob.glob(osj(run_informations["tmp_analysis_folder"], "*", "*.vcf.gz")) 
 
     for vcf_file in vcf_files:
         run = vcf_file.split("/")[-2]
@@ -49,88 +78,26 @@ def convert_vcf_parquet(run_informations, args):
             run_dict[run] = previous_value
         else:
             run_dict[run] = [vcf_name]
+        shutil.move(vcf_file, run_informations["tmp_analysis_folder"])
+    
+    vcf_files = glob.glob(osj(run_informations["tmp_analysis_folder"], "*.vcf.gz"))
+    
+    with Pool(int(threads)) as pool:
+        pool.map(minimize_worker_unpacker, [(vcf_file, run_informations,divide_memory(threads, memory)
+) for vcf_file in vcf_files])
 
-    for vcf_file in vcf_files:
-        subprocess.call(
-            ["rsync", "-rvt", vcf_file, run_informations["tmp_analysis_folder"]],
-            universal_newlines=True,
-        )
+    # lockfile = osj(
+    #     run_informations["tmp_analysis_folder"],
+    #     "dejavu_lock_" + run_informations["run_platform_application"],
+    # )
 
-    if "dejavu" in args:
-        for vcf_file in glob.glob(
-            osj(run_informations["tmp_analysis_folder"], "*.vcf.gz")
-        ):
-            howard_processing.unmerge_vcf(vcf_file, run_informations)
+    # if os.path.isfile(lockfile):
+    #     while os.path.isfile(lockfile):
+    #         sleep(60)
 
-    for vcf_file in glob.glob(osj(run_informations["tmp_analysis_folder"], "*.vcf.gz")):
-        tmp_output = osj(os.path.dirname(vcf_file), "tmp_" + os.path.basename(vcf_file))
-        cmd = ["bcftools", "annotate", "-x", "INFO", vcf_file]
-        with open(tmp_output, "w") as writefile:
-            subprocess.call(cmd, universal_newlines=True, stdout=writefile)
-        os.remove(vcf_file)
-        os.rename(tmp_output, vcf_file)
-
-        exact_time = time.time() + 7200
-        local_time = time.localtime(exact_time)
-        actual_time = time.strftime("%H%M%S", local_time)
-        start = actual_time
-        container_name = f"VANNOT_dejavu_{start}_{run_informations["run_name"]}_{os.path.basename(vcf_file).split(".")[0]}"
-        vcf_minimalize = osj(
-            run_informations["tmp_analysis_folder"],
-            f"{os.path.basename(vcf_file).split(".")[0]}.mini.parquet",
-        )
-        output_parquet = osj(
-            run_informations["tmp_analysis_folder"],
-            f"{os.path.basename(vcf_file).split(".")[0]}.parquet",
-        )
-
-        launch_minimalize_arguments = [
-            "minimalize",
-            "--input",
-            vcf_file,
-            "--output",
-            vcf_minimalize,
-            "--minimalize_info",
-            "--minimalize_samples",
-            "--threads",
-            threads,
-            "--memory",
-            memory,
-        ]
-        launch_parquet_arguments = [
-            "process",
-            "--input",
-            vcf_minimalize,
-            "--output",
-            output_parquet,
-            "--calculations=BARCODE",
-            "--explode_infos",
-            "--explode_infos_fields=barcode",
-            '--query=SELECT "#CHROM", POS, ID, REF, ALT, QUAL, FILTER, INFO, barcode FROM variants',
-            "--threads",
-            threads,
-            "--memory",
-            memory,
-        ]
-        log.info("Minimalizing vcfs")
-        howard_launcher.launch(container_name, launch_minimalize_arguments)
-        log.info("Converting to parquet format")
-        howard_launcher.launch(container_name, launch_parquet_arguments)
-        os.remove(vcf_minimalize)
-        os.remove(vcf_minimalize + ".hdr")
-
-    lockfile = osj(
-        run_informations["tmp_analysis_folder"],
-        "dejavu_lock_" + run_informations["run_platform_application"],
-    )
-
-    if os.path.isfile(lockfile):
-        while os.path.isfile(lockfile):
-            sleep(60)
-
-    if not os.path.isfile(lockfile):
-        with open(lockfile, "w") as write_file:
-            pass
+    # if not os.path.isfile(lockfile):
+    #     with open(lockfile, "w") as write_file:
+    #         pass
 
     for key, values in run_dict.items():
         for sample_name in values:
@@ -157,8 +124,69 @@ def convert_vcf_parquet(run_informations, args):
                 universal_newlines=True,
             )
 
-    # shutil.rmtree(run_informations["tmp_analysis_folder"])
+    shutil.rmtree(run_informations["tmp_analysis_folder"])
 
+
+def minimize_worker_unpacker(args):
+    vcf_file, run_informations, memory = args
+    minimize_worker(vcf_file, run_informations, memory)
+
+def minimize_worker(vcf_file: str, run_informations: dict, memory: str) -> None:
+    tmp_output = osj(os.path.dirname(vcf_file), "tmp_" + os.path.basename(vcf_file))
+    cmd = ["bcftools", "annotate", "-x", "INFO", vcf_file]
+    with open(tmp_output, "w") as writefile:
+        subprocess.call(cmd, universal_newlines=True, stdout=writefile)
+    os.remove(vcf_file)
+    os.rename(tmp_output, vcf_file)
+
+    exact_time = time.time() + 7200
+    local_time = time.localtime(exact_time)
+    actual_time = time.strftime("%H%M%S", local_time)
+    start = actual_time
+    container_name = f"VANNOT_dejavu_{start}_{run_informations["run_name"]}_{os.path.basename(vcf_file).split(".")[0]}"
+    vcf_minimalize = osj(
+        run_informations["tmp_analysis_folder"],
+        f"{os.path.basename(vcf_file).split(".")[0]}.mini.parquet",
+    )
+    output_parquet = osj(
+        run_informations["tmp_analysis_folder"],
+        f"{os.path.basename(vcf_file).split(".")[0]}.parquet",
+    )
+
+    launch_minimalize_arguments = [
+        "minimalize",
+        "--input",
+        vcf_file,
+        "--output",
+        vcf_minimalize,
+        "--minimalize_info",
+        "--minimalize_samples",
+        "--threads",
+        "1",
+        "--memory",
+        memory,
+    ]
+    launch_parquet_arguments = [
+        "process",
+        "--input",
+        vcf_minimalize,
+        "--output",
+        output_parquet,
+        "--calculations=BARCODE",
+        "--explode_infos",
+        "--explode_infos_fields=barcode",
+        '--query=SELECT "#CHROM", POS, ID, REF, ALT, QUAL, FILTER, INFO, barcode FROM variants',
+        "--threads",
+        "1",
+        "--memory",
+        memory,
+    ]
+    log.info("Minimalizing vcfs")
+    howard_launcher.launch(container_name, launch_minimalize_arguments)
+    log.info("Converting to parquet format")
+    howard_launcher.launch(container_name, launch_parquet_arguments)
+    # os.remove(vcf_minimalize)
+    # os.remove(vcf_minimalize + ".hdr")
 
 def calculate_dejavu(run_informations):
     threads = commons.get_threads("threads_dejavu")
@@ -208,7 +236,6 @@ def calculate_dejavu(run_informations):
             osj(run_informations["parquet_db_project_folder"], "*", "*", "*.parquet")
         )
     )
-
     log.info("Calculating new frequencies")
     allelecount = "ALLELECOUNT"
     hetcount = "HETCOUNT"
@@ -252,4 +279,5 @@ def calculate_dejavu(run_informations):
         writefile.write(
             f"#CHROM\tPOS\tREF\tALT\t{allelecount}\t{hetcount}\t{homcount}\t{allelefreq}\n"
         )
-    os.remove(lockfile)
+    if os.path.isfile(lockfile):
+        os.remove(lockfile)
